Add test scenes for bonus and final messages
Rearrange the test scenes so that the three scenes for testing level
transitions (`test_next_level`, `test_bonus`, and `test_final_level`)
are together, with the final level at the end.

Add a `Level` to `test_alligator` so that ENTER can move to the next
test scene. Remove the text stating that ENTER does not work.