Add an experimental target level
I do not know if I will end up using this, so put it in a new
"experimental" folder so as not to clutter up the rest of my directory.

Inspired by Runner 2's bonus levels, add a simple level where the player
is trying to jump onto a bullseye.

A lot more work is needed to flesh this out:
- The `target` needs to detect where the player lands on it
- The target needs a smaller bullseye that is harder to hit
- `Player` needs to properly handle colliding with something other than
  a `TileMapLayer`
- Something needs to detect when the player misses the target entirely