Replace workaround with call to `call_group`
In fa862fb4, `call_group` was not working, so I manually iterated
through the group members and called a method.

The reason it did not work was because I forgot to mark the function to
call on each group member as `#[func]`. This annotation is required for
Godot to find it[^1].

[^1]: https://godot-rust.github.io/book/register/functions.html?highlight=func#methods