WIP: Provide the `Main` entry point into the game
The level progression is currently stored in a linked list. While this
was easy to set up, it becomes harder to maintain. It also makes adding
in bonus levels, which should advance to the same level the player would
have reached had they missed the bonus and should be reusable, trickier.

This is the first step toward switching to a model with a top level node
that keeps track of global game state. It has an array of levels in
order, making it easier to modify the level progression. It can be used
in the future for other things, like a score that persists between
levels, or a user interface.

Add a `Main` scene, now the starting scene, which stores the list of
levels, editable in the editor. For now, include the first few scenes:
- title screen
- tutorial
- first level
This is enough to verify the basic behavior.

Adjust input handling. Since `Main` now handles the level progression,
it checks the input to determine when it is time to switch. This
simplifies the `Tutorial` code, which no longer needs to check for input
and switch to a hardcoded first level. In `MessageScreen`, consume input
when it should be ignored. After that, let `Main` handle it. In
`Player`, consume input while playing. Do not consume input during the
tutorial, when it should be used to skip to the next scene, or after
completing a level, for the same reason.

TODO: Add more levels and move to them appropriately
TODO: Use the same model for the test scenes