`Tutorial`: reset manually instead of using `SceneTree`
Using the `SceneTree` to reload made sense when the tutorial was the
main scene. In preparation for having a separate scene that cycles
through levels appropriately, avoid using the `SceneTree`.

Reset the player using `PlayerInfo`. Otherwise, when resetting from
the long jump, we lose the direction, and the player's next jump jumps
off screen to the left.