Remove `spawn_many_frogs`
This was interesting, but I do not have any plans to add it to the
actual game. `swarm` already covers the case where the player controls
multiple frogs at once. And this feature was broken by ongoing changes
to how input is handled. Remove it now to avoid trying to accommodate
it; it can be re-implemented later if truly desired.