Revert "Simplify `tutorial`'s button presses"
This reverts commit 1c8c93123f9809f71d4330babceb36e40e3fbc98.

Although the simplicity is nice, relying on the "jump" action means that
the player can interfere with the tutorial by continuing to hold "jump"
from the title screen. It is possible this can be worked around, but the
workarounds are not obvious, whereas this is more explicit about its
intent.

Conflicts:
    rust/src/tutorial.rs