Turn off most log statements by default
These `godot_print` statements were very helpful in getting jumping and
landing to work properly, but now they are less frequently necessary.
Leave them in, just in case they are helpful for bugs discovered in the
future, but turn them off by default.

Leave some logs (position and jump strength) as always-on, since they in
particular are useful for isolating the conditions for reproducing a
bug.

Add a new `#[export]`ed boolean to `Player` to toggle log statements.

Add `log.rs`, with a macro that conditionally calls `godot_print`.