Program the flies to travel based on their environment
Instead of requiring an `AnimationPlayer` for each fly, make a fly
travel in a straight line until they either collide with the "world" or
reach the edge of the platform they are above. Set most flies as "self-
directed", but turn this option off in the bonus level `swarm`.

Use `Direction` to choose the direction the fly is moving. Also flip the
fly in the editor if it is facing left.

Use a separate `Area2D` to make the fly turn around a little bit sooner.
Update collision masks. Now `Fly` itself (used by both `fly.tscn` and
`bonus_fly.tscn`) only detects collisions with the "player" layer, so
its `on_body_entered` no longer needs to check the groups. Similarly,
the new `Area2D` only detects the world.

In `swarm`, remove a fly that is hidden behind a frog and was
immediately eaten upon level start. Make the flies on the right face
left for symmetry.

In `vertical_level`, move the bonus fly so it hovers along a branch.
Adjust another fly so that its new flight path hints at the non-bonus
side of the vine.

TODO: Share scene tree between `fly.tscn` and `bonus_fly.tscn`?