Prevent using "NEXT" to skip past `test_bonus`
When the test level was originally written, pressing "NEXT" (ENTER)
would exit the test scenes, since it did not have a next level. This
changed in 3cc604087bb7c9fa52b97df07e2adf55aee7cf8d, which introduced
`test_main`. At that point, "NEXT"/ENTER skipped to the next scene,
defeating the goal of the test scene. But it also triggered a weird bug;
pressing "jump" in the *next* scene triggered the bonus level.

Disable "NEXT"/ENTER in this scene with a new rust script. This ensures
that the scene plays out as intended and avoids the bug. Update the
level description to note that ENTER does not work here.