Fix landing directly in the alligator's "eat" area
It is technically possible for the player to land in the "eat" area
- i.e. the `Area2D` named "EatArea2D" - at the same time as the
"OpenJawArea2D" and even the "FocusArea2D". The callbacks may be called
in any order. As previously written, this can result in bugs such as
the alligator never opening their mouth or never eating the player
(until they leave the "OpenJawArea2D" and return).

This is most easily noticeable in the `test_alligator` scene, but it
could be possible in other scenes with a very fast moving player, and/or
perhaps if we modify the various `Area2D`s later. Even though it may be
hard/impossible to reproduce in actual gameplay, fixing this makes the
test scene more helpful.

If the player enters both "open jaw" and "eat" simultaneously, open the
jaw for one frame and then close it the next.