Handle changing scenes using signals
The prior method of having `Main` interpret the "jump" presses as time
to advance to the next scene had the goal of avoiding `Main` having to
dynamically determine the type of the scene it loaded to connect its
signals. But that system resulted in less clear code. For example,
`MessageScreen` needed to set "jump" as handled when it didn't do
anything with it, and leave it unhandled for `Main` to advance to the
next scene.

Make `MessageScreen` and `Tutorial` send signals to advance to the next
scene on appropriate "jump" actions for clearer code.